Move around pkg files and setup for pytest
Restructure the project so deepcite is a package, and it is installed
in editable mode when requirements.txt is installed. This includes
adding a setup.py file.

Additionally, move main.py to test_basic.py so it becomes a quick
sanity check.
