+++ .travis.yaml
@@ -5,3 +5,4 @@ install:
   - pip install -r requirements.txt
 script:
   - pytest --pep8
+  - pytest
