+++ deepcite/compare.py
@@ -1,4 +1,4 @@
-from utils import *
+from .utils import *
 
 
 # returns the closest sections of text and their respective links,
